Generalize LR scheduler (#2345)
* start base scheduler

* fixes

* reorg

* add LearningRateScheduler base

* update custom LR schedulers

* fixes and update tests

* reorg tests

* fixes

* fix trainer

* fix docs

* fix trainer test

* make wrappers private

* improve docstring

* add momentum schedulers

* fix formatting for docs

* revert PyTorch LR scheduler behavior
