+++ allennlp/training/momentum_schedulers/__init__.py
@@ -1,0 +1,2 @@
+from allennlp.training.momentum_schedulers.momentum_scheduler import MomentumScheduler
+from allennlp.training.momentum_schedulers.inverted_triangular import InvertedTriangular
