+++ doc/api/allennlp.training.learning_rate_schedulers.rst
@@ -5,3 +5,23 @@ allennlp.training.learning_rate_schedulers
    :members:
    :undoc-members:
    :show-inheritance:
+
+.. automodule:: allennlp.training.learning_rate_schedulers.learning_rate_scheduler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.learning_rate_schedulers.cosine
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.learning_rate_schedulers.noam
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.learning_rate_schedulers.slanted_triangular
+   :members:
+   :undoc-members:
+   :show-inheritance:
