+++ doc/api/allennlp.training.momentum_schedulers.rst
@@ -1,0 +1,17 @@
+allennlp.training.momentum_schedulers
+==========================================
+
+.. automodule:: allennlp.training.momentum_schedulers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.momentum_schedulers.momentum_scheduler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.training.momentum_schedulers.inverted_triangular
+   :members:
+   :undoc-members:
+   :show-inheritance:
