+++ doc/api/allennlp.training.rst
@@ -7,7 +7,9 @@ for training AllenNLP models.
 .. toctree::
 
    allennlp.training.checkpointer
+   allennlp.training.scheduler
    allennlp.training.learning_rate_schedulers
+   allennlp.training.momentum_schedulers
    allennlp.training.metric_tracker
    allennlp.training.metrics
    allennlp.training.moving_average
