+++ doc/api/allennlp.training.scheduler.rst
@@ -1,0 +1,7 @@
+allennlp.training.scheduler
+==========================================
+
+.. automodule:: allennlp.training.scheduler
+   :members:
+   :undoc-members:
+   :show-inheritance:
